chore: organize and clean code
