+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-import App from './app'
+import App from './containers/app'
 import reportWebVitals from "./reportWebVitals";
 import "tachyons";
